requirements.txt
# Service principal authentication via env variables
python-dotenv==0.15.0

# Spark
pyspark==3.2.0
delta-spark==1.1.0